Allow multiple open entry streams at the same time
New entries can already be added while a previous entry stream is not
yet finished. The additional entries will be queued and will be written
once the previous entry is complete.

The queuing only happens if the user of the library adds more entries
before the callback function of the previous entry() call completed.
This is a usage pattern that wasn't supported previously, entry() did
throw in such a case. Existing users should not experience any additional
queueing.