Added a callback to finalize()
It is called once all data has been written and the generated tar stream
is finalized.

finalize() might not immediately finish the generated stream, and
multiple entry streams might still be pending to be read. It is convenient
to have a single callback that signals when all those pending actions are
complete.